Make projection validator generic
